+++ game-frontend/src/app/modules/configuration/types/configuration.type.ts
@@ -1,0 +1,14 @@
+
+/**
+ *
+ *
+ * @author Kevin Guanche Darias <kevin@kevinguanchedarias.com>
+ * @since 0.7.4
+ * @export
+ * @interface Configuration
+ */
+export interface Configuration<T> {
+    name: string;
+    displayName: string;
+    value: T;
+}